requirements.txt
coverage
Django
Faker
djangorestframework
djangorestframework-simplejwt
django-cors-headers